refactor(event): group event
