+++ src/main/java/sleppynavigators/studyupbackend/application/group/GroupService.java
@@ -11,9 +11,9 @@ import org.springframework.transaction.annotation.Transactional;
 import sleppynavigators.studyupbackend.domain.challenge.Challenge;
 import sleppynavigators.studyupbackend.domain.challenge.Task;
 import sleppynavigators.studyupbackend.domain.chat.Bot;
-import sleppynavigators.studyupbackend.domain.event.GroupCreateEvent;
-import sleppynavigators.studyupbackend.domain.event.UserJoinEvent;
-import sleppynavigators.studyupbackend.domain.event.UserLeaveEvent;
+import sleppynavigators.studyupbackend.domain.event.group.GroupCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.group.UserJoinEvent;
+import sleppynavigators.studyupbackend.domain.event.group.UserLeaveEvent;
 import sleppynavigators.studyupbackend.domain.group.Group;
 import sleppynavigators.studyupbackend.domain.group.GroupMember;
 import sleppynavigators.studyupbackend.domain.group.invitation.GroupInvitation;
