+++ src/main/java/sleppynavigators/studyupbackend/domain/chat/generator/action/ChallengeCreateChatActionListGenerator.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.stereotype.Component;
 import sleppynavigators.studyupbackend.domain.chat.action.ChatAction;
 import sleppynavigators.studyupbackend.domain.chat.action.ViewChallengeDetailChatAction;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCreateEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 
 @Component
