+++ src/main/java/sleppynavigators/studyupbackend/domain/chat/generator/message/ChallengeCompleteSystemMessageGenerator.java
@@ -1,7 +1,7 @@
 package sleppynavigators.studyupbackend.domain.chat.generator.message;
 
 import org.springframework.stereotype.Component;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCompleteEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCompleteEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 
 @Component
