+++ src/main/java/sleppynavigators/studyupbackend/domain/chat/generator/message/ChallengeCreateSystemMessageGenerator.java
@@ -1,7 +1,7 @@
 package sleppynavigators.studyupbackend.domain.chat.generator.message;
 
 import org.springframework.stereotype.Component;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCreateEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 
 @Component
