+++ src/main/java/sleppynavigators/studyupbackend/domain/event/challenge/ChallengeCompleteEvent.java
@@ -1,4 +1,8 @@
-package sleppynavigators.studyupbackend.domain.event;
+package sleppynavigators.studyupbackend.domain.event.challenge;
+
+import sleppynavigators.studyupbackend.domain.event.EventType;
+import sleppynavigators.studyupbackend.domain.event.PersonalNotificationEvent;
+import sleppynavigators.studyupbackend.domain.event.SystemMessageEvent;
 
 public record ChallengeCompleteEvent(
         String userName, String challengeName, Long groupId, Long challengeId, Long challengerId, Double percentage)
