+++ src/main/java/sleppynavigators/studyupbackend/domain/event/challenge/TaskCertifyEvent.java
@@ -1,4 +1,8 @@
-package sleppynavigators.studyupbackend.domain.event;
+package sleppynavigators.studyupbackend.domain.event.challenge;
+
+import sleppynavigators.studyupbackend.domain.event.EventType;
+import sleppynavigators.studyupbackend.domain.event.GroupNotificationEvent;
+import sleppynavigators.studyupbackend.domain.event.SystemMessageEvent;
 
 public record TaskCertifyEvent(
         String userName, String challengeName, String taskName, Long groupId, Long challengeId, Long taskId
