+++ src/main/java/sleppynavigators/studyupbackend/domain/event/group/UserLeaveEvent.java
@@ -1,4 +1,7 @@
-package sleppynavigators.studyupbackend.domain.event;
+package sleppynavigators.studyupbackend.domain.event.group;
+
+import sleppynavigators.studyupbackend.domain.event.EventType;
+import sleppynavigators.studyupbackend.domain.event.SystemMessageEvent;
 
 public record UserLeaveEvent(String userName, Long groupId) implements SystemMessageEvent {
