+++ src/test/java/sleppynavigators/studyupbackend/application/challenge/ChallengeServiceTest.java
@@ -23,9 +23,9 @@ import sleppynavigators.studyupbackend.common.support.GroupSupport;
 import sleppynavigators.studyupbackend.common.support.UserSupport;
 import sleppynavigators.studyupbackend.domain.challenge.Challenge;
 import sleppynavigators.studyupbackend.domain.chat.Bot;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCancelEvent;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCreateEvent;
-import sleppynavigators.studyupbackend.domain.event.TaskCertifyEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCancelEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.TaskCertifyEvent;
 import sleppynavigators.studyupbackend.domain.group.Group;
 import sleppynavigators.studyupbackend.domain.user.User;
 import sleppynavigators.studyupbackend.infrastructure.user.UserRepository;
