+++ src/test/java/sleppynavigators/studyupbackend/application/group/GroupServiceTest.java
@@ -17,9 +17,9 @@ import sleppynavigators.studyupbackend.common.support.BotSupport;
 import sleppynavigators.studyupbackend.common.support.GroupSupport;
 import sleppynavigators.studyupbackend.common.support.UserSupport;
 import sleppynavigators.studyupbackend.domain.chat.Bot;
-import sleppynavigators.studyupbackend.domain.event.GroupCreateEvent;
-import sleppynavigators.studyupbackend.domain.event.UserJoinEvent;
-import sleppynavigators.studyupbackend.domain.event.UserLeaveEvent;
+import sleppynavigators.studyupbackend.domain.event.group.GroupCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.group.UserJoinEvent;
+import sleppynavigators.studyupbackend.domain.event.group.UserLeaveEvent;
 import sleppynavigators.studyupbackend.domain.group.Group;
 import sleppynavigators.studyupbackend.domain.group.invitation.GroupInvitation;
 import sleppynavigators.studyupbackend.domain.user.User;
