+++ src/test/java/sleppynavigators/studyupbackend/domain/chat/generator/action/TaskFailChatActionListGeneratorTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.Test;
 import sleppynavigators.studyupbackend.domain.chat.action.ChatAction;
 import sleppynavigators.studyupbackend.domain.chat.action.ChatActionType;
 import sleppynavigators.studyupbackend.domain.event.EventType;
-import sleppynavigators.studyupbackend.domain.event.TaskFailEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.TaskFailEvent;
 
 public class TaskFailChatActionListGeneratorTest {
