+++ src/test/java/sleppynavigators/studyupbackend/domain/chat/generator/message/ChallengeCompleteSystemMessageGeneratorTest.java
@@ -3,7 +3,7 @@ package sleppynavigators.studyupbackend.domain.chat.generator.message;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCompleteEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCompleteEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 
 class ChallengeCompleteSystemMessageGeneratorTest {
