+++ src/test/java/sleppynavigators/studyupbackend/domain/chat/generator/message/ChallengeCreateSystemMessageGeneratorTest.java
@@ -3,7 +3,7 @@ package sleppynavigators.studyupbackend.domain.chat.generator.message;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
-import sleppynavigators.studyupbackend.domain.event.ChallengeCreateEvent;
+import sleppynavigators.studyupbackend.domain.event.challenge.ChallengeCreateEvent;
 import sleppynavigators.studyupbackend.domain.event.EventType;
 
 class ChallengeCreateSystemMessageGeneratorTest {
