+++ src/test/java/sleppynavigators/studyupbackend/domain/chat/generator/message/UserJoinSystemMessageGeneratorTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.Test;
 import sleppynavigators.studyupbackend.domain.event.EventType;
-import sleppynavigators.studyupbackend.domain.event.UserJoinEvent;
+import sleppynavigators.studyupbackend.domain.event.group.UserJoinEvent;
 
 class UserJoinSystemMessageGeneratorTest {
